Separated tests into single files
Renamed LedgerIndex to CommittedIndex
Prepared indexer to insert/delete outputs on acceptance
