+++ pkg/indexer/types.go
@@ -22,23 +22,23 @@ type LedgerUpdate struct {
 }
 
 type LedgerOutput struct {
-	OutputID  iotago.OutputID
-	Output    iotago.Output
-	CreatedAt iotago.SlotIndex
-	SpentAt   iotago.SlotIndex
+	OutputID iotago.OutputID
+	Output   iotago.Output
+	BookedAt iotago.SlotIndex
+	SpentAt  iotago.SlotIndex
 }
 
 type Status struct {
 	ID              uint `gorm:"primaryKey;notnull"`
-	LedgerIndex     iotago.SlotIndex
+	CommittedIndex  iotago.SlotIndex
 	NetworkName     string
 	DatabaseVersion uint32
 }
 
 type queryResult struct {
-	OutputID    []byte
-	Cursor      string
-	LedgerIndex iotago.SlotIndex
+	OutputID       []byte
+	Cursor         string
+	CommittedIndex iotago.SlotIndex
 }
 
 type queryResults []queryResult
@@ -54,11 +54,11 @@ func (q queryResults) IDs() iotago.OutputIDs {
 
 //nolint:revive // better be explicit here
 type IndexerResult struct {
-	OutputIDs   iotago.OutputIDs
-	LedgerIndex iotago.SlotIndex
-	PageSize    uint32
-	Cursor      *string
-	Error       error
+	OutputIDs      iotago.OutputIDs
+	CommittedIndex iotago.SlotIndex
+	PageSize       uint32
+	Cursor         *string
+	Error          error
 }
 
 func errorResult(err error) *IndexerResult {
@@ -141,10 +141,10 @@ func (i *Indexer) combineOutputIDFilteredQueries(queries []*gorm.DB, pageSize ui
 }
 
 func (i *Indexer) resultsForQuery(query *gorm.DB, pageSize uint32) *IndexerResult {
-	// This combines the query with a second query that checks for the current ledger_index.
+	// This combines the query with a second query that checks for the current committed_index.
 	// This way we do not need to lock anything and we know the index matches the results.
-	ledgerIndexQuery := i.db.Model(&Status{}).Select("ledger_index")
-	joinedQuery := i.db.Table("(?) as results, (?) as status", query, ledgerIndexQuery)
+	committedIndexQuery := i.db.Model(&Status{}).Select("committed_index")
+	joinedQuery := i.db.Table("(?) as results, (?) as status", query, committedIndexQuery)
 
 	var results queryResults
 
@@ -155,11 +155,11 @@ func (i *Indexer) resultsForQuery(query *gorm.DB, pageSize uint32) *IndexerResul
 
 	var ledgerIndex iotago.SlotIndex
 	if len(results) > 0 {
-		ledgerIndex = results[0].LedgerIndex
+		ledgerIndex = results[0].CommittedIndex
 	} else {
 		// Since we got no results for the query, return the current ledger index
 		if status, err := i.Status(); err == nil {
-			ledgerIndex = status.LedgerIndex
+			ledgerIndex = status.CommittedIndex
 		}
 	}
 
@@ -172,10 +172,10 @@ func (i *Indexer) resultsForQuery(query *gorm.DB, pageSize uint32) *IndexerResul
 	}
 
 	return &IndexerResult{
-		OutputIDs:   results.IDs(),
-		LedgerIndex: ledgerIndex,
-		PageSize:    pageSize,
-		Cursor:      nextCursor,
-		Error:       nil,
+		OutputIDs:      results.IDs(),
+		CommittedIndex: ledgerIndex,
+		PageSize:       pageSize,
+		Cursor:         nextCursor,
+		Error:          nil,
 	}
 }
