+++ pkg/server/routes.go
@@ -764,10 +764,10 @@ func outputsResponseFromResult(result *indexer.IndexerResult) (*outputsResponse,
 	}
 
 	return &outputsResponse{
-		LedgerIndex: result.LedgerIndex,
-		PageSize:    result.PageSize,
-		Cursor:      cursor,
-		Items:       result.OutputIDs.ToHex(),
+		CommittedIndex: result.CommittedIndex,
+		PageSize:       result.PageSize,
+		Cursor:         cursor,
+		Items:          result.OutputIDs.ToHex(),
 	}, nil
 }
